requirements.txt
# Deep learning and image processing
torch
torchvision
numpy
matplotlib
Pillow
streamlit
langchain
langchain-openai
langchain-community
openai
pypdf
qdrant-client
sentence-transformers
dotenv
groq
